Extract dialogs to hooks
